Added worker threads option.
Now with the command line option -w you can specify how many worker threads to use. This is a simple first pass at handling better concurrency. ebb_rails defaults to 1 worker.

+++ benchmark/application.rb
@@ -35,7 +35,7 @@ class SimpleApp
       status = 200
     
     elsif commands.include?('wait')
-      n = commands.last.to_i
+      n = commands.last.to_f
       raise "wait called with n <= 0" if n <= 0
       wait(n)
       body = "waited about #{n} seconds"
